fix(core): update corel contexts to match new lint rules
