+++ packages/sanity/src/_singletons/context/ReleasesTableContext.ts
@@ -1,0 +1,11 @@
+import {createContext} from 'sanity/_createContext'
+
+import type {TableContextValue} from '../../core/releases/components/Table/TableProvider'
+
+/**
+ * @internal
+ */
+export const TableContext = createContext<TableContextValue | null>(
+  'sanity/_singletons/context/releases-table',
+  null,
+)
