+++ packages/sanity/src/_singletons/index.ts
@@ -1,5 +1,6 @@
 export * from './context/ActiveWorkspaceMatcherContext'
 export * from './context/AddonDatasetContext'
+export * from './context/BundlesMetadataContext'
 export * from './context/CalendarContext'
 export * from './context/ChangeIndicatorTrackerContexts'
 export * from './context/ColorSchemeSetValueContext'
@@ -43,6 +44,7 @@ export * from './context/PresenceTrackerContexts'
 export * from './context/PreviewCardContext'
 export * from './context/ReferenceInputOptionsContext'
 export * from './context/ReferenceItemRefContext'
+export * from './context/ReleasesTableContext'
 export * from './context/ResourceCacheContext'
 export * from './context/ReviewChangesContext'
 export * from './context/RouterContext'
@@ -66,5 +68,3 @@ export * from './context/VirtualizerScrollInstanceContext'
 export * from './context/WorkspaceContext'
 export * from './context/WorkspacesContext'
 export * from './context/ZIndexContext'
-export * from './core/releases/BundlesMetadataContext'
-export * from './core/releases/ReleasesTableContext'
