+++ packages/sanity/src/core/releases/components/Table/TableHeader.tsx
@@ -1,7 +1,7 @@
 import {ArrowDownIcon, ArrowUpIcon, SearchIcon} from '@sanity/icons'
 import {Button, type ButtonProps, Card, Flex, Stack, TextInput} from '@sanity/ui'
-import {useTableContext} from 'sanity/_singletons'
 
+import {useTableContext} from './TableProvider'
 import {type HeaderProps, type TableHeaderProps} from './types'
 
 const SortHeaderButton = ({header, text}: ButtonProps & HeaderProps) => {
