+++ packages/sanity/src/core/releases/components/Table/TableProvider.tsx
@@ -1,9 +1,19 @@
-import {type ComponentType, type PropsWithChildren, useCallback, useState} from 'react'
+import {type ComponentType, type PropsWithChildren, useCallback, useContext, useState} from 'react'
 import {TableContext} from 'sanity/_singletons'
 
 export interface TableSort {
   column: string
   direction: 'asc' | 'desc'
+}
+
+/**
+ * @internal
+ */
+export interface TableContextValue {
+  searchTerm: string | null
+  setSearchTerm: (searchTerm: string) => void
+  sort: TableSort | null
+  setSortColumn: (column: string) => void
 }
 
 /**
@@ -29,4 +39,15 @@ export const TableProvider: ComponentType<PropsWithChildren & {defaultSort?: Tab
   const contextValue = {searchTerm, setSearchTerm, sort, setSortColumn}
 
   return <TableContext.Provider value={contextValue}>{children}</TableContext.Provider>
+}
+
+/**
+ * @internal
+ */
+export const useTableContext = (): TableContextValue => {
+  const context = useContext(TableContext)
+  if (!context) {
+    throw new Error('useTableContext must be used within a TableProvider')
+  }
+  return context
 }
